# api-account-service/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Account Service API
  description: |
    **Microservicio de Gestión de Cuentas y Movimientos**

    Este microservicio maneja toda la lógica transaccional:
    - Gestión de cuentas bancarias (CRUD completo)
    - Registro y procesamiento de movimientos (débitos/créditos)
    - Control de saldos y validaciones transaccionales
    - Generación de reportes de estado de cuenta
    - Procesamiento de eventos asíncronos desde Customer Service

    **Arquitectura:**
    - Spring Boot + WebFlux (Reactivo)
    - Arquitectura Hexagonal (Ports & Adapters)
    - Event-Driven Architecture
    - Contract-First approach
    - Circuit Breaker para resiliencia
    - Cache distribuido para optimización

    **Consideraciones de rendimiento:**
    - Transacciones ACID garantizadas
    - Optimistic locking para concurrencia
    - Rate limiting para protección
    - Monitoreo de métricas transaccionales

    ---
    **Senior Level** - Technical Assessment `2025`
  contact:
    name: Darwin Josue Pilaloa Zea
    email: dpilaloazea@gmail.com
  version: 1.0.0
servers:
  - url: http://localhost:8082/api/v1
    description: Local Development
  - url: https://dev-account-api.com/api/v1
    description: Development Environment
  - url: https://account-api.com/api/v1
    description: Production Environment
...
security:
  - Authorization: []
paths:
  /accounts:
    get:
      tags:
        - Accounts
      summary: Obtiene listado paginado de cuentas
      description: |
        Retorna una lista paginada de cuentas bancarias.
        Soporta filtrado por cliente, tipo y estado.
      operationId: getAccounts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: customerId
          in: query
          description: Filtrar por ID de cliente
          required: false
          schema:
            $ref: '#/components/schemas/UUID'
        - name: accountType
          in: query
          description: Filtrar por tipo de cuenta
          required: false
          schema:
            $ref: '#/components/schemas/AccountType'
        - name: state
          in: query
          description: Filtrar por estado activo/inactivo
          required: false
          schema:
            type: boolean
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
      responses:
        '200':
          description: Listado de cuentas recuperado exitosamente
          headers:
            X-Total-Count:
              description: Total de registros disponibles
              schema:
                type: integer
            X-Page-Number:
              description: Número de página actual
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountPageResponse'
              examples:
                SuccessfulResponse:
                  $ref: '#/components/examples/AccountsPageExample'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Accounts
      summary: Crea una nueva cuenta bancaria
      description: |
        Crea una nueva cuenta bancaria asociada a un cliente.

        **Validaciones de negocio:**
        - El cliente debe existir y estar activo (validación asíncrona)
        - El número de cuenta debe ser único
        - El saldo inicial no puede ser negativo
        - Máximo 5 cuentas activas por cliente
      operationId: createAccount
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        description: Datos de la cuenta a crear
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreateRequest'
            examples:
              SavingsAccount:
                $ref: '#/components/examples/CreateSavingsAccountExample'
              CheckingAccount:
                $ref: '#/components/examples/CreateCheckingAccountExample'
      responses:
        '201':
          description: Cuenta creada exitosamente
          headers:
            Location:
              description: URI del recurso creado
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Cliente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflicto - Número de cuenta duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /accounts/{accountNumber}:
    get:
      tags:
        - Accounts
      summary: Obtiene una cuenta por número
      description: Retorna los datos completos de una cuenta específica
      operationId: getAccountByNumber
      parameters:
        - $ref: '#/components/parameters/AccountNumberParam'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
      responses:
        '200':
          description: Cuenta encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags:
        - Accounts
      summary: Actualiza el estado de una cuenta
      description: |
        Permite cambiar el estado activo/inactivo de una cuenta.

        **Validaciones:**
        - No se puede desactivar una cuenta con saldo positivo
        - No se puede activar una cuenta si el cliente está inactivo
      operationId: patchAccountState
      parameters:
        - $ref: '#/components/parameters/AccountNumberParam'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountStatePatchRequest'
      responses:
        '200':
          description: Estado actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Accounts
      summary: Elimina una cuenta (soft delete)
      description: |
        Realiza una eliminación lógica de la cuenta.

        **Validaciones:**
        - La cuenta debe tener saldo cero
        - No debe tener movimientos pendientes
      operationId: deleteAccount
      parameters:
        - $ref: '#/components/parameters/AccountNumberParam'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
      responses:
        '204':
          description: Cuenta eliminada exitosamente
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflicto - Cuenta con saldo o movimientos pendientes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /accounts/{accountNumber}/balance:
    get:
      tags:
        - Accounts
      summary: Consulta el saldo actual de una cuenta
      description: |
        Retorna el saldo disponible actualizado de la cuenta.
        Endpoint optimizado para consultas frecuentes (con cache).
      operationId: getAccountBalance
      parameters:
        - $ref: '#/components/parameters/AccountNumberParam'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
      responses:
        '200':
          description: Saldo recuperado exitosamente
          headers:
            Cache-Control:
              schema:
                type: string
              description: Directivas de cache
              example: "max-age=60"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    Authorization:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT en el header Authorization.
        Formato: Authorization: Bearer {token}
  parameters:
    AccountNumberParam:
      name: accountNumber
      in: path
      description: Número de cuenta bancaria
      required: true
      schema:
        $ref: '#/components/schemas/AccountNumber'
...
    PageParam:
      name: page
      in: query
      description: Número de página (0-indexed)
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      description: Tamaño de página
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      description: Campo de ordenamiento
      required: false
      schema:
        type: string
        default: "createdAt,desc"
    XRequestId:
      name: X-Request-Id
      in: header
      description: ID único de la petición para trazabilidad
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
    XCorrelationId:
      name: X-Correlation-Id
      in: header
      description: ID de correlación para trazar operaciones entre servicios
      required: false
      schema:
        $ref: '#/components/schemas/UUID'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: Clave de idempotencia para operaciones críticas
      required: false
      schema:
        $ref: '#/components/schemas/UUID'
  schemas:
    UUID:
      type: string
      format: uuid
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
      minLength: 36
      maxLength: 36
      example: "550e8400-e29b-41d4-a716-446655440000"
    AccountNumber:
      type: integer
      format: int64
      description: Número único de cuenta bancaria
      minimum: 100000
      maximum: 9999999999
      example: 478758
    AccountType:
      type: string
      description: Tipo de cuenta bancaria
      enum:
        - AHORRO
        - CORRIENTE
      example: "AHORRO"
...
    Money:
      type: number
      format: double
      description: Valor monetario (siempre positivo para transacciones)
      minimum: 0
      multipleOf: 0.01
      example: 1500.50
    AccountCreateRequest:
      type: object
      required:
        - accountType
        - initialBalance
        - customerId
      properties:
        accountNumber:
          $ref: '#/components/schemas/AccountNumber'
          description: Número de cuenta (opcional, se autogenera si no se provee)
        accountType:
          $ref: '#/components/schemas/AccountType'
        initialBalance:
          $ref: '#/components/schemas/Money'
          description: Saldo inicial de la cuenta
        customerId:
          $ref: '#/components/schemas/UUID'
          description: ID del cliente propietario
        state:
          type: boolean
          description: Estado inicial de la cuenta
          default: true
    AccountStatePatchRequest:
      type: object
      required:
        - state
      properties:
        state:
          type: boolean
          description: Nuevo estado de la cuenta
        reason:
          type: string
          description: Razón del cambio de estado
          minLength: 5
          maxLength: 200
          example: "Solicitud del cliente"
    AccountResponse:
      type: object
      properties:
        accountNumber:
          $ref: '#/components/schemas/AccountNumber'
        accountType:
          $ref: '#/components/schemas/AccountType'
        balance:
          $ref: '#/components/schemas/Money'
        state:
          type: boolean
          example: true
        customerId:
          $ref: '#/components/schemas/UUID'
        customerName:
          type: string
          description: Nombre del cliente (desnormalizado para performance)
          example: "Jose Lema"
        createdAt:
          type: string
          format: date-time
          example: "2025-11-06T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-11-06T10:30:00.000Z"
    AccountPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AccountResponse'
        page:
          $ref: '#/components/schemas/PageMetadata'
    BalanceResponse:
      type: object
      properties:
        accountNumber:
          $ref: '#/components/schemas/AccountNumber'
        balance:
          $ref: '#/components/schemas/Money'
        availableBalance:
          $ref: '#/components/schemas/Money'
          description: Saldo disponible (considerando retenciones)
        timestamp:
          type: string
          format: date-time
          description: Timestamp de la consulta
          example: "2025-11-06T10:30:00.000Z"
...
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          example: 20
        number:
          type: integer
          example: 0
        totalElements:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2025-11-06T10:30:00.000Z"
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Validation failed"
        path:
          type: string
          example: "/api/v1/accounts"
        traceId:
          $ref: '#/components/schemas/UUID'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    ValidationError:
      type: object
      properties:
        field:
          type: string
          example: "amount"
        message:
          type: string
          example: "must be greater than 0"
        rejectedValue:
          example: -100
        additionalInfo:
          type: object
          additionalProperties: true
...
  responses:
    BadRequest:
      description: Petición incorrecta - Error de validación
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: No autorizado - Token inválido o expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Entidad no procesable - Error de lógica de negocio
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: Servicio no disponible o degradado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  examples:
    CreateSavingsAccountExample:
      value:
        accountNumber: 478758
        accountType: "AHORRO"
        initialBalance: 2000.00
        customerId: "550e8400-e29b-41d4-a716-446655440000"
        state: true
    CreateCheckingAccountExample:
      value:
        accountType: "CORRIENTE"
        initialBalance: 1000.00
        customerId: "550e8400-e29b-41d4-a716-446655440000"
        state: true
    AccountsPageExample:
      value:
        content:
          - accountNumber: 478758
            accountType: "AHORRO"
            balance: 2000.00
            state: true
            customerId: "550e8400-e29b-41d4-a716-446655440000"
            customerName: "Jose Lema"
            createdAt: "2025-11-06T10:30:00.000Z"
            updatedAt: "2025-11-06T10:30:00.000Z"
          - accountNumber: 225487
            accountType: "CORRIENTE"
            balance: 700.00
            state: true
            customerId: "660e8400-e29b-41d4-a716-446655440001"
            customerName: "Marianela Montalvo"
            createdAt: "2025-11-06T10:30:00.000Z"
            updatedAt: "2025-11-06T10:30:00.000Z"
        page:
          size: 20
          number: 0
          totalElements: 2
          totalPages: 1
...